python/agents/machine-learning-engineering/pyproject.toml
[build-system]
requires = [ "poetry-core>=2.0.0,<3.0.0",]
build-backend = "poetry.core.masonry.api"

[project]
name = "machine_learning_engineering"
version = "0.1"
description = "MLE-STAR: A novel machine learning engineering (MLE) agent to automate the implementation of ML models."
readme = "README.md"
requires-python = ">= 3.12, < 4.0"
dependencies = [
    "google-adk (>=1.5.0,<2.0.0)", 
    "google-genai (>=1.9.0,<2.0.0)", 
    "pydantic (>=2.10.6,<3.0.0)", 
    "python-dotenv (>=1.0.1,<2.0.0)", 
    "google-cloud-aiplatform (>=1.93,<2.0)", 
    "numpy (>=2.2.3,<3.0.0)", 
    "pandas (>=2.3.1,<3.0.0)", 
    "scikit-learn (>=1.7.1,<2.0.0)", 
    "torch (>=2.7.1,<3.0.0)",
    "absl-py (>=2.2.1)"
]
[[project.authors]]
name = "Jiefeng Chen"
email = "jiefengc@google.com"

[[project.authors]]
name = "Raj Sinha"
email = "sinharaj@google.com"

[[project.authors]]
name = "Jaehyun Nam"
email = "jaehyunnam@google.com"

[[project.authors]]
name = "Jinsung Yoon"
email = "jinsungyoon@google.com"

[project.license]
text = "Apache License 2.0"

[project.optional-dependencies]
dev = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
